requirements.txt
flask
sqlalchemy
psycopg2-binary
flask_sqlalchemy
python-dotenv
